requirements.txt
requests
ipython
python-dotenv
